refactor: utilize rsc for complaint page
