+++ src/app/parliament/complaint/page.tsx
@@ -1,122 +1,28 @@
-'use client';
+import { FetchReturn, GetQuery, PaginatedParliamentTickets } from '@/lib/types';
+import { STATUS_OPTIONS, TIME_OPTIONS } from '@/lib/constants';
+import ParliamentFilter from '../(components)/ParliamentFilter';
+import ParliamentSearch from '../(components)/ParliamentSearch';
+import ParliamentTable from '../(components)/ParliamentTable';
+import { Skeleton } from '@/components/ui/skeleton';
+import { Suspense } from 'react';
+import { getComplaintTickets } from './api';
 
-import { formatDateToString, subtractMonths, subtractWeeks } from '@/lib/utils';
-import { useEffect, useState } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { BsSearch } from 'react-icons/bs';
-import ParliamentEmptyState from '../(components)/ParliamentEmptyState';
-import ParliamentSkeletonCard from '../(components)/ParliamentSkeletonCard';
-import ParliamentTableBottom from '../(components)/ParliamentTableBottom';
-import ParliamentTicketingPagination from '../(components)/ParliamentTicketingPagination';
-import { ParliamentTickets } from '@/lib/types';
-import { getComplaintTickets } from './actions';
-import useQueryHandler from '@/lib/hooks/useQueryHandler';
-
-export default function Complaint() {
-  //pagination state
-  const [currentPage, setCurrentPage] = useState(1);
-  const [rowsPerCurrentPage, setRowsPerCurrentPage] = useState(10);
-
-  //to find status in url query
-  const searchParams = useSearchParams();
-  const statusParams = searchParams.get('status');
-
-  //filter state
-  const [filterStatus, setfilterStatus] = useState(
-    statusParams ? `&status=${statusParams}` : ''
-  );
-
-  const [filterTime, setfilterTime] = useState('');
-  const [search, setSearch] = useState('');
-  const filter = `${filterStatus}${filterTime}`;
-
-  useEffect(() => {
-    setCurrentPage(1);
-  }, [filterStatus, filterTime, search]);
-
-  // Define a state variable to avoid spamming enter key
-  const [searchCooldown, setSearchCooldown] = useState(false);
-
-  //search function
-  const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const words = e.target.value;
-    if (words !== '') {
-      setSearch(`&search=${words}`);
-    } else {
-      setSearch('');
-    }
+export default async function Complaint({
+  searchParams,
+}: {
+  searchParams: {
+    period_end?: string;
+    status?: string;
+    search?: string;
+    page?: number;
+    'per-page'?: number;
   };
-
-  //to avoid spamming enter button when searching
-  useEffect(() => {
-    const keyDownHandler = (e: KeyboardEvent) => {
-      if (e.key === 'Enter' && search !== '' && !searchCooldown) {
-        setSearchCooldown(true);
-        e.preventDefault();
-        // refetch();
-      }
-      if (searchCooldown) {
-        setTimeout(() => {
-          setSearchCooldown(false);
-        }, 3000);
-      }
-    };
-
-    document.addEventListener('keydown', keyDownHandler);
-
-    return () => {
-      document.removeEventListener('keydown', keyDownHandler);
-    };
-  }, [search, searchCooldown]);
-
-  //navigate to id page
-  const router = useRouter();
-  const handleRowClick = (itemId: string) => {
-    router.push(`/parliament/aid/${itemId}`);
-  };
-
-  //set time filter
-  const getCurrentDate = () => {
-    const currentDate = new Date();
-    return `&&period_end=${formatDateToString(currentDate)}`;
-  };
-
-  const getPastWeekDate = () => {
-    const pastWeek = subtractWeeks(new Date(), 1);
-    return `&&period_end=${formatDateToString(pastWeek)}`;
-  };
-
-  const getPastMonthDate = () => {
-    const pastMonth = subtractMonths(new Date(), 1);
-    return `&&period_end=${formatDateToString(pastMonth)}`;
-  };
-
-  const {
-    data: complaintTicketData,
-    state: getComplaintTicketsState,
-    triggerQuery: triggerGetComplaintTickets,
-  } = useQueryHandler({
-    query: () =>
-      getComplaintTickets({
-        page: currentPage,
-        perPage: rowsPerCurrentPage,
-        filter,
-        search,
-      }),
-    deps: [currentPage, rowsPerCurrentPage, filter],
-  });
-  //set debounce for search filter
-  useEffect(() => {
-    // Set a new timeout to refetch data after 3 seconds if words are not empty
-    const debounceTimeout = setTimeout(() => {
-      triggerGetComplaintTickets();
-    }, 2000);
-
-    return () => {
-      clearTimeout(debounceTimeout);
-    };
-    //eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [search]);
+}) {
+  const complaintTickets: FetchReturn<GetQuery<PaginatedParliamentTickets>> =
+    await getComplaintTickets({
+      perPage: searchParams['per-page'],
+      ...searchParams,
+    });
 
   return (
     <div>
@@ -126,190 +32,32 @@ export default function Complaint() {
         <div className='w-full'>
           <div className='flex items-center justify-between'>
             <div className='flex gap-4'>
-              <select
-                value={filterStatus}
-                onChange={(e) => setfilterStatus(e.target.value)}
-                className='border text-body1 border-linePrimary w-[240px] h-[48px] px-2 py-0.5 rounded focus:outline-none focus:border-linePrimary'
-              >
-                <option value=''>All status</option>
-                <option value='&status=COMPLETED'>Completed</option>
-                <option value='&status=PENDING'>Pending</option>
-                <option value='&status=REJECTED'>Rejected</option>
-              </select>
-              <select
-                onChange={(e) => setfilterTime(e.target.value)}
-                className='border text-body1 border-linePrimary w-[240px] h-[48px] px-2 py-0.5 rounded focus:outline-none focus:border-linePrimary'
-              >
-                <option value=''>All time</option>
-                <option value={getCurrentDate()}>Today</option>
-                <option value={getPastWeekDate()}>Past week</option>
-                <option value={getPastMonthDate()}>Past month</option>
-              </select>
+              <Suspense fallback={<Skeleton className='w-full h-[6.5rem]' />}>
+                <ParliamentFilter
+                  paramKey='status'
+                  options={STATUS_OPTIONS}
+                />
+              </Suspense>
+              <Suspense fallback={<Skeleton className='w-full h-[6.5rem]' />}>
+                <ParliamentFilter
+                  paramKey='period_end'
+                  options={TIME_OPTIONS}
+                />
+              </Suspense>
             </div>
-            <div className='flex items-center bg-white rounded border border-linePrimary px-3 py-2'>
-              <BsSearch className='text-gray-400' />
-              <input
-                type='text'
-                placeholder='Search ID or title'
-                className='ml-2 px-2 py-0.5 text-body1 text-textPrimary rounded border-none outline-none shadow-none active:border-none'
-                onChange={(e) => handleSearch(e)}
-              />
-            </div>
+            <ParliamentSearch />
           </div>
           <div className='mt-8 flex flex-col overflow-x-hidden'>
             <div className=' sm:-mx-6 lg:-mx-8'>
               <div className='inline-block min-w-full py-2 align-middle md:px-6 lg:px-8'>
-                <table className='min-w-full '>
-                  <thead>
-                    <tr className='bg-baseBg'>
-                      <th
-                        scope='col'
-                        className='py-3.5 pl-4 pr-3 text-left text-body1  text-textSecondary sm:pl-6 md:pl-3'
-                      >
-                        ID
-                      </th>
-                      <th
-                        scope='col'
-                        className='py-3.5 px-3 text-left text-body1  text-textSecondary'
-                      >
-                        Requested on
-                      </th>
-                      <th
-                        scope='col'
-                        className='py-3.5 px-3 text-left text-body1  text-textSecondary'
-                      >
-                        Title
-                      </th>
-                      <th
-                        scope='col'
-                        className='py-3.5 px-3 text-left text-body1  text-textSecondary w-[150px]'
-                      >
-                        Status
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody className='divide-y divide-lineSecondary'>
-                    {/* loading state */}
-                    {getComplaintTicketsState === 'pending' && (
-                      <tr>
-                        <td className='whitespace-nowrap py-4 pl-4 pr-3 text-body1 text-textPrimary sm:pl-6 md:pl-3'>
-                          <ParliamentSkeletonCard />
-                        </td>
-                        <td className='whitespace-nowrap py-4 px-3 text-body1 text-textPrimary'>
-                          <ParliamentSkeletonCard />
-                        </td>
-                        <td className='whitespace-nowrap py-4 px-3 text-body1 text-textPrimary'>
-                          <ParliamentSkeletonCard />
-                        </td>
-                        <td className='whitespace-nowrap py-4 px-3 text-body1 text-textPrimary flex items-center gap-2'>
-                          <ParliamentSkeletonCard />
-                        </td>
-                      </tr>
-                    )}
-
-                    {/* normal state */}
-                    {getComplaintTicketsState === 'resolved' &&
-                      complaintTicketData.ok &&
-                      complaintTicketData?.data.data.data.map(
-                        (item: ParliamentTickets, index: number) => (
-                          <tr
-                            onClick={() => handleRowClick(`${item.id}`)}
-                            key={index}
-                            className='cursor-pointer'
-                          >
-                            <td className='whitespace-nowrap py-4 pl-4 pr-3 text-body1 text-textPrimary sm:pl-6 md:pl-3'>
-                              {item.id}
-                            </td>
-                            <td className='whitespace-nowrap py-4 px-3 text-body1 text-textPrimary'>
-                              {
-                                new Date(item.created_at)
-                                  .toString()
-                                  .split('GMT')[0]
-                              }
-                            </td>
-                            {/* <Link href="/"> */}
-                            <td className='whitespace-nowrap py-4 px-3 text-body1 text-textPrimary'>
-                              {item.title}
-                            </td>
-                            {/* </Link> */}
-
-                            <td className='whitespace-nowrap py-4 px-3 text-body1 text-textPrimary flex items-center gap-2'>
-                              <span>
-                                {item.ticket_status === 'COMPLETED' && (
-                                  <div className='w-[10px] h-[10px] rounded-full bg-positive'></div>
-                                )}
-                                {item.ticket_status === 'PENDING' && (
-                                  <div className='w-[10px] h-[10px] rounded-full bg-warning'></div>
-                                )}
-                                {item.ticket_status === 'AWAITING_REPLY' && (
-                                  <div className='w-[10px] h-[10px] rounded-full bg-waitingColor'></div>
-                                )}
-                                {item.ticket_status === 'REJECTED' && (
-                                  <div className='w-[10px] h-[10px] rounded-full bg-negative'></div>
-                                )}
-                              </span>
-                              <span>
-                                {item.ticket_status === 'AWAITING_REPLY'
-                                  ? 'Awaiting reply'
-                                  : item.ticket_status}
-                              </span>
-                            </td>
-                          </tr>
-                        )
-                      )}
-
-                    {/* error state */}
-                    {(getComplaintTicketsState === 'resolved' ||
-                      getComplaintTicketsState === 'error') &&
-                      ((complaintTicketData.ok &&
-                        complaintTicketData?.data.data.data.length === 0) ||
-                        !complaintTicketData.ok) && (
-                        <tr>
-                          <td colSpan={4}>
-                            <ParliamentEmptyState />
-                          </td>
-                        </tr>
-                      )}
-                  </tbody>
-                </table>
+                <Suspense fallback={<Skeleton className='w-full h-full' />}>
+                  <ParliamentTable complaintTicketData={complaintTickets} />
+                </Suspense>
               </div>
             </div>
           </div>
-          {/* PAGINATION */}
-          {getComplaintTicketsState === 'resolved' &&
-            complaintTicketData.ok &&
-            complaintTicketData?.data.data.data.length !== 0 && (
-              <ParliamentTicketingPagination
-                count={complaintTicketData?.data.data.meta.total}
-                lastPage={complaintTicketData?.data.data.meta.lastPage}
-                pages={currentPage}
-                rows={rowsPerCurrentPage}
-                setPages={setCurrentPage}
-                setRows={setRowsPerCurrentPage}
-                from={
-                  complaintTicketData?.data.data.meta.currentPage *
-                    rowsPerCurrentPage -
-                  rowsPerCurrentPage +
-                  1
-                }
-                to={
-                  complaintTicketData?.data.data.meta.currentPage ===
-                  complaintTicketData?.data.data.meta.lastPage
-                    ? complaintTicketData?.data.data.meta.total
-                    : complaintTicketData?.data.data.meta.currentPage *
-                      rowsPerCurrentPage
-                }
-              />
-            )}
-          {(getComplaintTicketsState === 'error' ||
-            getComplaintTicketsState === 'pending' ||
-            (complaintTicketData.ok &&
-              complaintTicketData.data.data.data.length === 0)) && (
-            <ParliamentTableBottom />
-          )}
         </div>
       </main>
     </div>
   );
 }
-
